requirements.txt
fastmcp
pyyaml
gitpython
